+++ 9_fdm_delete_access_policies.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 '''
-Copyright (c) 2019 Cisco and/or its affiliates.
+Copyright (c) 2020 Cisco and/or its affiliates.
 
 This software is licensed to you under the terms of the Cisco Sample
 Code License, Version 1.1 (the "License"). You may obtain a copy of the
@@ -27,6 +27,9 @@ import csv
 from pprint import pprint
 from requests.packages.urllib3.exceptions import InsecureRequestWarning
 requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
+from crayons import blue, green, white, red, yellow,magenta, cyan
+
+new_auth_token=['none']#as global variable in order to make it easily updatable 
 
 def yaml_load(filename):
 	fh = open(filename, "r")
@@ -53,11 +56,17 @@ def fdm_login(host,username,password,version):
 		raise Exception("Error logging in: {}".format(request.content))
 	try:
 		access_token = request.json()['access_token']
+		fa = open("token.txt", "w")
+		fa.write(access_token)
+		fa.close()	
+		new_auth_token[0]=access_token
+		print (green("Token = "+access_token))
+		print("Saved into token.txt file")		
 		return access_token
 	except:
 		raise
 
-def delete_access_policy_from_csv(host,token,file,parent_id,version):
+def delete_access_policy_from_csv(host,token,file,parent_id,version,username,password):
 	'''
 	Delete every service object from the csv file
 	'''
@@ -74,13 +83,39 @@ def delete_access_policy_from_csv(host,token,file,parent_id,version):
 			print(row[0]+' id = '+row[7])
 			try:
 				if row[0].find('NEW_')==0:
-					request = requests.delete("https://{}:{}/api/fdm/v{}/policy/accesspolicies/{}/accessrules/{}".format(host, FDM_PORT,version,parent_id,row[7]), headers=headers, verify=False)  					
+					request = requests.delete("https://{}:{}/api/fdm/v{}/policy/accesspolicies/{}/accessrules/{}".format(host, FDM_PORT,version,parent_id,row[7]), headers=headers, verify=False)  		
+					status_code = request.status_code
+					if status_code == 401: 
+						print(red("Auth Token invalid, Let\'s ask for a new one",bold=True))
+						fdm_login(host,username,password,version)
+						line_content = []
+						with open('token.txt') as inputfile:
+							for line in inputfile:
+								if line.strip()!="":	
+									line_content.append(line.strip())						
+						auth_token = line_content[0]
+						#headers["Authorization"]="Bearer {}".format(auth_token)	
+						headers = {
+							"Content-Type": "application/json",
+							"Accept": "application/json",
+							"Authorization":"Bearer {}".format(auth_token)
+						}			
+						request = requests.delete("https://{}:{}/api/fdm/v{}/policy/accesspolicies/{}/accessrules/{}".format(host, FDM_PORT,version,parent_id,row[7]), headers=headers, verify=False)
+						status_code = request.status_code	
+					resp = request.text
+					if status_code == 200 or status_code == 201 or status_code == 202 or status_code == 204:
+						print (green("Delete was successful...",bold=True))
+						#json_resp = json.loads(resp)
+						#print(json.dumps(json_resp,sort_keys=True,indent=4, separators=(',', ': ')))
+					else :
+						request.raise_for_status()
+						print (red("Error occurred in Delete --> "+resp+' Status Code = '+str(status_code)))						
 					print("Access Policy removed")
 			except:
 				raise			
 	return (1)	
 	
-def fdm_get(host,token,url,version):
+def fdm_get(host,token,url,version,username,password):
 	'''
 	generic GET request.
 	'''
@@ -91,7 +126,26 @@ def fdm_get(host,token,url,version):
 	}
 	try:
 
-		request = requests.get("https://{}:{}/api/fdm/v{}{}?limit=100".format(host, FDM_PORT,version,url),verify=False, headers=headers)
+		request = requests.get("https://{}:{}/api/fdm/v{}{}?limit=100".format(host, FDM_PORT,version,url),verify=False, headers=headers)		
+		status_code = request.status_code
+		if status_code == 401: 
+			print(red("Auth Token invalid, Let\'s ask for a new one",bold=True))
+			fdm_login(host,username,password,version)
+			line_content = []
+			with open('token.txt') as inputfile:
+				for line in inputfile:
+					if line.strip()!="":	
+						line_content.append(line.strip())						
+			auth_token = line_content[0]
+			#headers["Authorization"]="Bearer {}".format(auth_token)	
+			headers = {
+				"Content-Type": "application/json",
+				"Accept": "application/json",
+				"Authorization":"Bearer {}".format(auth_token)
+			}			
+			request = requests.get("https://{}:{}/api/fdm/v{}{}?limit=100".format(host, FDM_PORT,version,url),verify=False, headers=headers)
+			status_code = request.status_code	
+						
 		return request.json()
 	except:
 		raise
@@ -111,14 +165,14 @@ if __name__ == "__main__":
 	fa = open("token.txt", "r")
 	token = fa.readline()
 	fa.close()
-	#token = fdm_login(FDM_HOST,FDM_USER,FDM_PASSWORD) 
+	new_auth_token[0]=token
 	print()
 	print (" TOKEN :")
 	print(token)
 	print('======================================================================================================================================')	 
 	# STEP 1  Get the Policy ID , we need it as the parent ID for accessrules management	
 	api_url="/policy/accesspolicies"
-	accesspolicy = fdm_get(FDM_HOST,token,api_url,FDM_VERSION)
+	accesspolicy = fdm_get(FDM_HOST,token,api_url,FDM_VERSION,FDM_USER,FDM_PASSWORD)
 	#print(json.dumps(accesspolicy,indent=4,sort_keys=True))
 	data=accesspolicy['items']
 	for entry in data:
@@ -128,4 +182,5 @@ if __name__ == "__main__":
 	
 	file="access_policies.txt"
 	print("OBJECTS TO DELETE :")
-	delete_access_policy_from_csv(FDM_HOST,token,file,PARENT_ID,FDM_VERSION)
+	token=new_auth_token[0]
+	delete_access_policy_from_csv(FDM_HOST,token,file,PARENT_ID,FDM_VERSION,FDM_USER,FDM_PASSWORD)
